requirements.txt
asgiref==3.7.2
Django==4.2.5
pyasn1==0.5.0
pycryptodome==3.19.0
sqlparse==0.4.4
typing_extensions==4.8.0
